requirements.txt
streamlit
pandas
google-generativeai>=0.8.3
folium
streamlit-folium